feat: Export client only ClerkProvider
