+++ src/lib/client/index.ts
@@ -1,0 +1,1 @@
+export { default as ClerkProvider } from './ClerkProvider.svelte';
