+++ src/lib/types.ts
@@ -1,3 +1,4 @@
+import type { LoadClerkJsScriptOptions } from '@clerk/shared';
 import type { Clerk, ClerkOptions, ClientResource, Without } from '@clerk/types';
 import type { Snippet } from 'svelte';
 
@@ -12,6 +13,11 @@ export interface BrowserClerk extends HeadlessBrowserClerk {
 }
 
 export type PropsWithChildren<T> = T & { children?: Snippet };
+
+export type ClerkProviderProps = Omit<LoadClerkJsScriptOptions, 'publishableKey'> & {
+	publishableKey?: string;
+	children: Snippet;
+};
 
 declare global {
 	interface Window {
